+++ .github/workflows/test.yml
@@ -23,4 +23,4 @@ jobs:
         python -m pip install --upgrade pip
         pip install picard
     - name: Test plugins
-      run: python test.py
+      run: python -m unittest discover -v
